requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.3
Django==5.2.6
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-environ==0.12.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.3.0
python-decouple==3.8
reportlab==4.4.4
sqlparse==0.5.3